2. 后端最终版 v1.0/API_Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-dotenv
httpx
PyPDF2
aiofiles
pdf2image
python-pptx
Pillow
passlib[bcrypt]
python-jose[cryptography]
paddlepaddle
paddleocr
alibabacloud_dysmsapi20170525==2.0.24
